+++ src/engine/flow/main/declared-cost-params.ts
@@ -10,6 +10,7 @@ export function declaredCostParamsToDyn(costParams?: AbilityCostParams): Record<
     'sleepChar',
     'stunChar',
     'removeFromHand',
+    'revealFromHand',
     'sceneToDeckBottom',
     'removeAreaToDeckBottom',
     'partnerAreaRemove',